cybersec_soc_env/pyproject.toml
[project]
name = "cybersec-soc-env"
version = "0.1.0"
description = "SOC analyst RL environment with procedural topology, alert noise, and business impact scoring"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
keywords = ["reinforcement-learning", "cybersecurity", "openenv", "rl-environment"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Security",
]

dependencies = [
    "openenv-core",
    "networkx>=3.2",
    "numpy>=1.26",
    "fastapi>=0.110",
    "uvicorn>=0.29",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["cybersec_soc_env"]
